Add element selection and get variables
an array of html elements is produced from first parameter
and an object of variables is produced from second
parameter and/or defaults.

It is so far compatible with greensocks splittext but
also extends slightly.

The position property is not limited to absolute and
relative but will also accept static, fixed, initial
 or inherit.  There probably isn't going to be much
need for these but there might be cases.  It will
also accept null which won't set the position property
on the elements.